django_server/requirements.txt
coverage==4.5.2
Django==2.1.7
django-filter==2.1.0
django-nose==1.4.6
djangorestframework==3.9.1
Markdown==3.0.1
nose==1.3.7
psycopg2==2.7.7
pytz==2018.9
PyYAML==3.13
